feat(daemon): Phase 6 Commit C — adaptive idle-demote + per-drive min_tier + shard.ttl tracing + index decomposition
Phase 6 Commit C of the memory-tiering work
(`docs/refactor/memory-tiering-implementation-plan.md` §3 Phase 6).
Wires the daemon.toml-driven adaptive TTL ladder (Commit A formulas
+ Commit B parser) into the live demote path, threads the operator's
per-drive `min_tier` overrides through the controller, emits
structured `shard.ttl` tracing on every demote evaluation, and
decomposes `index/mod.rs` into nine cohesive sibling modules so the
file-size policy is back green without an exception.  Plan tasks
6.3 / 6.4 / 6.6 / 6.7 / 6.8 closed; 6.1 + 6.2 + 6.5 re-validated
end-to-end through the new wiring.

# Phase 6 Commit C wiring

* `crate::run_daemon` now resolves `daemon.toml` via a new
  `load_daemon_config()` helper (platform-default path; missing →
  Phase-3 defaults; malformed → startup error with line/column).
* `IndexManager::new` consolidated to take `Arc<Config>` explicitly;
  the default-config sugar is the test surface only and every test
  site (15 migrated) now passes
  `Arc::new(crate::config::Config::default())`.
* `index::transitions::demote_idle_shards` now calls a new
  `evaluate_idle_demote(shard, now_ms, &config)` helper which:
    1. Samples the live `decay_ema_qpm` from the per-drive
       `DriveStats` (× 60 conversion added in `cache::shard`).
    2. Builds per-shard `TierThresholds` via
       `build_thresholds(rate_qpm, &config.tiers)` (uses `hot_ttl`
       / `warm_ttl` / `parked_ttl` from `cache::policy`).
    3. Calls `next_state_for_idle_with_thresholds` with adaptive
       thresholds (replaces the static `next_state_for_idle` shim
       — which moves into `mod tests` since Commit C made it
       production-unreachable).
    4. Applies a per-drive `min_tier` floor via
       `min_tier_for_drive` (case-insensitive lookup, `"C"` and
       `"C:"` both supported).
    5. Emits `tracing::debug!(target: "shard.ttl", ...)` with
       `drive`, `from`, `to`/`proposed`, `chosen_ttl_sec`,
       `rate_qpm`, and a structured `reason` (`idle-demote` /
       `min-tier-clamp` / `below-ttl`).
* Env-var override contract preserved end-to-end:
  `TiersConfig::default()` and `ShardsConfig::default()` now call
  the env-var-aware `hot_to_warm_idle_secs()` /
  `warm_to_parked_idle_secs()` / `parked_to_cold_idle_secs()` /
  `usn_refresh_interval_secs()` getters, so `UFFS_*_IDLE_SECS` /
  `UFFS_USN_REFRESH_INTERVAL_SECS` flow through `Config` when no
  `daemon.toml` is present.
* Phase-3 thin wrappers (`next_state_for_idle`,
  `TierThresholds::from_static_policy`) moved into `mod tests`
  since Commit C made them production-unreachable.  The production
  `cache::policy` surface now exposes only the adaptive helpers.
* Stale `#[expect(dead_code, reason = "Phase 6 Commit C wires
  this...")]` attributes removed from `hot_ttl` / `warm_ttl` /
  `parked_ttl` (and a handful of `Config` parser methods) — Commit
  C is what wired them, so the expectations were unfulfilled.
* Type renamed `DaemonConfig` → `Config` to avoid collision with
  the existing `cli::DaemonConfig` runtime-args wrapper that
  `lib.rs::run_daemon` already takes; the daemon.toml schema lives
  at `crate::config::Config` (idiomatic `crate::module::Type`).

# `index/mod.rs` decomposition (Phase 6 prep)

`crates/uffs-daemon/src/index/mod.rs` carved from 2 413 LOC down
to ~348 LOC (file-size policy back green without an exception)
by extracting nine cohesive sibling modules:

* `constructors.rs` (225 LOC) — `IndexManager::new` +
  `LifecycleHooks` builder bundle (cuts every constructor's
  arity from 8 to 4).
* `dispatch.rs` (506 LOC) — `IndexManager::record_search_dispatch`
  + `ensure_warm_for_dispatch` + the per-letter single-flight
  promote machinery (PR-e dedup).
* `hotload.rs` (387 LOC) — drive-add hot-load loop
  (`add_drive_with_loader_*`).
* `info.rs` (196 LOC) — `IndexManager::status` snapshot +
  `drives` RPC handler.
* `loading.rs` (565 LOC) — `load_compact_caches` + offline MFT
  scan path + atomic `add_drive` / `replace_drive`.
* `refresh.rs` (223 LOC) — Phase 5 (#95) USN refresh helpers
  separated from the transition controller.
* `stats.rs` (147 LOC) — `IndexManager::record_query` +
  tier-aware bookkeeping.
* `test_helpers.rs` (130 LOC) — `_for_test` constructors + the
  `FixedBodyLoader` family of test-only fakes.
* `transitions.rs` (712 LOC) — Phase 3 Commit D
  `demote_idle_shards`, Phase 5 task 5.6 `cascade_demote_one_step`,
  Phase 5 #95 `refresh_usn_for_warm_shards` + Phase 6 Commit C's
  `evaluate_idle_demote` adaptive helpers.

`scripts/ci/file_size_exceptions.txt` no longer carries
`crates/uffs-daemon/src/index/mod.rs`; `bash
scripts/ci/check_file_size_policy.sh` is green without it.

# Tests

5 new tests in `index::transitions::commit_c_tests` covering plan
tasks 6.4 / 6.6 / 6.7:

* `per_drive_min_tier_clamps_proposed_demote_below_floor`
* `min_tier_for_drive_case_insensitive_and_colon_optional`
* `build_thresholds_at_zero_rate_matches_base_secs`
* `build_thresholds_at_rate_20_extends_hot_edge`
* `chosen_ttl_for_state_picks_outgoing_edge`

The `defaults_match_phase3_static_behavior` test in `config::tests`
now uses a `const _: () = { ... }` static-pin block to anchor the
constants to the policy layer, so it remains deterministic under
any `OnceLock`-cached env state.

# Verification

* `just lint-prod` — clean.
* `just check-windows` — clean (xwin x86_64-pc-windows-msvc).
* `just check-all-targets` — Mac + Windows green; Linux Docker
  N/A on this dev host (picked up in PR-Fast CI).
* `cargo nextest run --workspace --all-features --all-targets` —
  1619 / 1619 passing, 14 skipped (live-host integration ipc tests).
